css-align-3_dev/html/place-self-shorthand-003.htm

<!DOCTYPE html>
<html><head><title>CSS Box Alignment: place-self shorthand - initial value</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez">
<link href="http://www.w3.org/TR/css3-align/#place-self-property" rel="help">
<meta content="Check that place-self's 'initial' value expands to 'align-self' and 'justify-self'." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
    #test {
        align-self: start;
        justify-self: end;
    }
</style>
</head><body><div id="log"></div>
<div id="test"></div>
<script>
    var div = document.getElementById("test");
    div.setAttribute("style", "place-self: initial");
    var style = getComputedStyle(div);

    test(function() {
        assert_equals(style.getPropertyValue("place-self"),
            "normal normal", "place-self resolved value");
    }, "Check place-self: initial - resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("align-self"),
            "normal", "place-self specified value for align-self");
    }, "Check place-self: initial - align-self resolved value");

    test(function() {
        assert_equals(style.getPropertyValue("justify-self"),
            "normal", "place-self specified value for justify-self");
    }, "Check place-self: initial - justify-self resolved value");
</script>
</body></html>
